fix: algoritmos mejorados funcionando correctamente

+++ BruteF.h
@@ -10,17 +10,17 @@ struct Point {
     int x, y;
 };
 
-float distance(const Point& p1, const Point& p2) {
+double distance(const Point& p1, const Point& p2) {
     return sqrtf((p1.x - p2.x) * (p1.x - p2.x) +
                  (p1.y - p2.y) * (p1.y - p2.y));
 }
 
-float bruteForce(const vector<Point>& P) {
-    float minDist = numeric_limits<float>::max();
+double bruteForce(const vector<Point>& P) {
+    double minDist = numeric_limits<double>::max();
     int n = P.size();
     for (int i = 0; i < n; ++i) {
         for (int j = i + 1; j < n; ++j) {
-            float d = distance(P[i], P[j]);
+            double d = distance(P[i], P[j]);
             if (d < minDist)
                 minDist = d;
         }
